fix(messaging): address PR review comments
- Extract shared fetchSignalRToken helper into lib/signalr-token.ts;
  chat-hub and notification-hub both import from there to avoid drift
- Add staleTime: 60_000 to useConversations — SignalR invalidations
  handle real-time freshness so aggressive background refetching is
  unnecessary now that the query is used globally for the nav badge
- Add aria-label ("Messages, N unread") to the Messages link in
  MainNav and MobileNav when unread count > 0
- Add aria-hidden and sr-only "Unread" text to the chat-list dot
  so assistive tech users get the unread signal
- Backfill seeker_last_read_at / helper_last_read_at to last_message_at
  in the migration so existing conversations don't show as unread on deploy
- Update sender's LastReadAt in SendAsync so sending a message never
  marks the conversation as unread for the sender
