+++ frontend/lib/chat-hub.ts
@@ -6,20 +6,7 @@ import { useQueryClient } from "@tanstack/react-query"
 
 import { conversationKeys } from "@/lib/api/conversations"
 import type { ApiMessage } from "@/lib/api/conversations"
-import { refreshAccessToken } from "@/lib/auth/token-bridge"
-
-async function fetchToken(): Promise<string> {
-  let res = await fetch("/api/auth/token")
-  if (res.status === 401) {
-    const refreshed = await refreshAccessToken()
-    if (refreshed) {
-      res = await fetch("/api/auth/token")
-    }
-  }
-  if (!res.ok) return ""
-  const data = (await res.json()) as { token: string }
-  return data.token
-}
+import { fetchSignalRToken } from "@/lib/signalr-token"
 
 export function useConversationHub(
   conversationId: string,
@@ -32,7 +19,7 @@ export function useConversationHub(
 
     const connection = new signalR.HubConnectionBuilder()
       .withUrl("/hubs/chat", {
-        accessTokenFactory: fetchToken,
+        accessTokenFactory: fetchSignalRToken,
       })
       .withAutomaticReconnect()
       .build()
